backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
psycopg2-binary>=2.9
gunicorn>=21.2
python-dotenv>=1.0
django-cors-headers>=4.3
pytest-django>=4.7
pytest-cov>=4.1
model_bakery>=1.17
celery>=5.3.6
redis>=5.0.1
django-redis>=5.4.0
djangorestframework-simplejwt>=5.3.1
pre-commit>=3.6.0
ruff>=0.1.13
mypy>=1.8.0
django-stubs>=4.2.7
python-json-logger>=2.0.7
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
pytest-mock>=3.11.1
factory-boy>=3.3.0
